+++ js/unix_formatting.js
@@ -13,17 +13,52 @@
  * Released under the MIT license
  *
  */
-/* global jQuery */
-(function($) {
-    if (!$.terminal) {
-        throw new Error('$.terminal is not defined');
+/* global jQuery, define, global, require, module */
+(function(factory) {
+    var root = typeof window !== 'undefined' ? window : global;
+    if (typeof define === 'function' && define.amd) {
+        // AMD. Register as an anonymous module.
+        // istanbul ignore next
+        define(['jquery', 'jquery.terminal'], factory);
+    } else if (typeof module === 'object' && module.exports) {
+        // Node/CommonJS
+        module.exports = function(root, jQuery) {
+            if (jQuery === undefined) {
+                // require('jQuery') returns a factory that requires window to
+                // build a jQuery instance, we normalize how we use modules
+                // that require this pattern but the window provided is a noop
+                // if it's defined (how jquery works)
+                if (typeof window !== 'undefined') {
+                    jQuery = require('jquery');
+                } else {
+                    jQuery = require('jquery')(root);
+                }
+            }
+            if (!jQuery.fn.terminal) {
+                if (typeof window !== 'undefined') {
+                    require('jquery.terminal');
+                } else {
+                    require('jquery.terminal')(jQuery);
+                }
+            }
+            factory(jQuery);
+            return jQuery;
+        };
+    } else {
+        // Browser
+        // istanbul ignore next
+        factory(root.jQuery);
     }
-
+})(function($) {
     $.terminal.defaults.unixFormattingEscapeBrackets = false;
+    // we match characters and html entities because command line escape brackets
+    // echo don't, when writing formatter always process html entitites so it work
+    // for cmd plugin as well for echo
     var chr = '[^\\x08]|[\\r\\n]{2}|&[^;]+;';
     var backspace_re = new RegExp('^(' + chr + ')?\\x08');
     var overtyping_re = new RegExp('^(?:(' + chr + ')?\\x08(_|\\1)|' +
                                    '(_)\\x08(' + chr + '))');
+    var new_line_re = /^(\r\n|\n\r|\r|\n)/;
     // ---------------------------------------------------------------------
     function length(string) {
         return $.terminal.length(string);
@@ -32,18 +67,19 @@
     // :: Replace overtyping (from man) formatting with terminal formatting
     // ---------------------------------------------------------------------
     $.terminal.overtyping = function overtyping(string, options) {
+        string = $.terminal.unescape_brackets(string);
         var settings = $.extend({
             position: 0
         }, options);
         var removed_chars = [];
         var new_position;
+        var char_count = 0;
+        var backspaces = [];
         function replace(string, position) {
-            // we match characters and html entities because command line escape brackets
-            // echo don't, when writing formatter always process html entitites so it work
-            // for cmd plugin as well for echo
             var result = '';
             var push = 0;
             var start;
+            char_count = 0;
             function correct_position(start, match, rep_string) {
                 // logic taken from $.terminal.tracking_replace
                 if (start < position) {
@@ -69,7 +105,7 @@
                 if (match) {
                     // we remove backspace and character or html entity before it
                     // but we keep it in removed array so we can put it back
-                    // when we have caritage return
+                    // when we have caritage return or line feed
                     if (match[1]) {
                         start = i - match[1].length + push;
                         removed_chars.push({
@@ -79,23 +115,46 @@
                         });
                         correct_position(start, match[0], '', 1);
                     }
-                    return result + partial.replace(backspace_re, '');
-                } else if (string[i] === '\r') {
-                    // if newline we need to add at the end all characters
-                    // removed by backspace
-                    if (removed_chars.length) {
-                        removed_chars = removed_chars.reduce(function(acc, char) {
-                            if (i > char.index) {
-                                correct_position(char.index, '', char.string, 2);
-                                result += char.string;
-                                return acc;
-                            }
-                            acc.push(char);
-                            return acc;
-                        }, []);
+                    if (char_count < 0) {
+                        char_count = 0;
                     }
-                    result += string[i];
+                    backspaces = backspaces.map(function(b) {
+                        return b - 1;
+                    });
+                    backspaces.push(start);
+                    return result + partial.replace(backspace_re, '');
+                } else if (partial.match(new_line_re)) {
+                    // if newline we need to add at the end all characters
+                    // removed by backspace but only if there are no more
+                    // other characters than backspaces added between
+                    // backspaces and newline
+                    if (removed_chars.length) {
+                        var chars = removed_chars;
+                        removed_chars = [];
+                        chars.reverse().forEach(function(char) {
+                            if (i > char.index) {
+                                if (--char_count <= 0) {
+                                    correct_position(char.index, '', char.string, 2);
+                                    result += char.string;
+                                }
+                            } else {
+                                removed_chars.unshift(char);
+                            }
+                        });
+                    }
+                    var m = partial.match(new_line_re);
+                    result += m[1];
+                    i += m[1].length - 1;
                 } else {
+                    if (backspaces.length) {
+                        var backspace = backspaces[0];
+                        if (i === backspace) {
+                            backspaces.shift();
+                        }
+                        if (i >= backspace) {
+                            char_count++;
+                        }
+                    }
                     if (removed_chars.length) {
                         // if we are in index of removed character we check if the
                         // character is the same it will be bold or if removed char
@@ -371,8 +430,9 @@
             }
             return [styles.join(''), color, background];
         }
-        var clear_line_re = /[^\r\n]+\r\x1B(?:&#91;|\[)K/g;
+        var clear_line_re = /[^\r\n]+\r\x1B\[K/g;
         return function from_ansi(input, options) {
+            input = $.terminal.unescape_brackets(input);
             var settings = $.extend({
                 unixFormattingEscapeBrackets: false,
                 position: 0
@@ -390,9 +450,10 @@
             var splitted = input.split(/(\x1B\[[0-9;]*[A-Za-z])/g);
             if (splitted.length === 1) {
                 if (settings.unixFormattingEscapeBrackets) {
-                    result = $.terminal.escape_brackets(input);
+                    result = $.terminal.escape_formatting(input);
+                } else {
+                    result = input;
                 }
-                result = input;
                 if (typeof options.position === 'number') {
                     return [result, new_position];
                 }
@@ -414,13 +475,13 @@
                     var start = match.index;
                     // logic taken from $.terminal.tracking_replace
                     if (start < position) {
-                        var last_index = start + $.terminal.length(match);
+                        var last_index = start + $.terminal.length(match[0]);
                         if (last_index < position) {
                             // It's after the replacement, move it
                             new_position = Math.max(
                                 0,
                                 new_position +
-                                    $.terminal.length(match)
+                                    $.terminal.length(match[0])
                             );
                         } else {
                             // It's *in* the replacement, put it just after
@@ -467,7 +528,7 @@
                             break;
                     }
                 } else if (settings.unixFormattingEscapeBrackets) {
-                    output.push($.terminal.escape_brackets(splitted[i]));
+                    output.push($.terminal.escape_formatting(splitted[i]));
                 } else {
                     output.push(splitted[i]);
                 }
@@ -485,4 +546,4 @@
 
     $.terminal.defaults.formatters.unshift($.terminal.overtyping);
     $.terminal.defaults.formatters.unshift($.terminal.from_ansi);
-})(jQuery);
+});